requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain
langchain-openai
